+++ CHANGELOG.md
@@ -108,6 +108,7 @@
 
 ### Chore & Maintenance
 
+* `[babel-jest]` [**BREAKING**] Always return object from transformer (TBD)
 * `[jest-jasmine2]` Simplify `Env.execute` and TreeProcessor to setup and clean
   resources for the top suite the same way as for all of the children suites
   ([#5885](https://github.com/facebook/jest/pull/5885))